refactor(woof): extract WOOF response logic into a separate utility function
