+++ packages/bruno-electron/src/ipc/network/index.js
@@ -37,6 +37,7 @@ const { cookiesStore } = require('../../store/cookies');
 const registerGrpcEventHandlers = require('./grpc-event-handlers');
 const { registerWsEventHandlers } = require('./ws-event-handlers');
 const { getCertsAndProxyConfig, buildCertsAndProxyConfig } = require('./cert-utils');
+const { buildWoofResponse } = require('../../utils/woof');
 const { buildFormUrlEncodedPayload, isFormData, extractBoundaryFromContentType } = require('@usebruno/common').utils;
 
 const ERROR_OCCURRED_WHILE_EXECUTING_REQUEST = 'Error occurred while executing the request!';
@@ -876,40 +877,7 @@ const registerNetworkIpc = (mainWindow) => {
 
     // Every good boy deserves a response.
     if (request.method && request.method.toUpperCase() === 'WOOF') {
-      const woofStart = Date.now();
-      const body = [
-        'Woof! Woof!',
-        '',
-        '       __',
-        '   (___()\'`;',
-        '   /,    /`',
-        '   \\\\"--\\\\',
-        '',
-        'Bruno fetched your request. Good human.'
-      ].join('\n');
-      const buffer = Buffer.from(body, 'utf-8');
-      return {
-        status: 200,
-        statusText: 'Woof!',
-        headers: {
-          'content-type': 'text/plain; charset=utf-8',
-          'x-good-boy': 'true',
-          'x-fetched-by': 'bruno'
-        },
-        data: body,
-        dataBuffer: buffer.toString('base64'),
-        size: buffer.byteLength,
-        duration: Date.now() - woofStart,
-        url: request.url,
-        timeline: [],
-        requestSent: {
-          url: request.url,
-          method: 'WOOF',
-          headers: request.headers,
-          data: null,
-          timestamp: woofStart
-        }
-      };
+      return buildWoofResponse(request);
     }
 
     request.__bruno__executionMode = 'standalone';
